WIP: Initial failed attempt at webpack
