+++ .gitignore
@@ -2,3 +2,6 @@ bower_components
 node_modules
 *.iml
 .idea
+bundle.js
+css/materialize.min.css
+css/starter-template.css
