--- bower.json
@@ -1,33 +1,0 @@
-{
-  "name": "c-quince",
-  "version": "0.0.1",
-  "homepage": "https://github.com/MallowWorld/c-quince",
-  "authors": [
-    "Stephen Riesenberg <sjohnr@hotmail.com>"
-  ],
-  "description": "A friendly avian robot used to teach kids how to write code, perhaps without realizing it.",
-  "main": "src/c-quince/c-quince.js",
-  "moduleType": [
-    "globals"
-  ],
-  "dependencies": {
-    "materialize-css": "~0.97.5",
-    "TweenJS": "~0.6.2",
-    "EaselJS": "~0.8.2"
-  },
-  "keywords": [
-    "animated",
-    "c-quince",
-    "kids",
-    "code"
-  ],
-  "license": "LGPL",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ],
-  "private": true
-}
