+++ demo.html
@@ -2,9 +2,9 @@
 <html>
 <head>
   <link href="http://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-  <link type="text/css" rel="stylesheet" href="bower_components/Materialize/dist/css/materialize.min.css" media="screen,projection"/>
-  <link type="text/css" rel="stylesheet" href="bower_components/Materialize/templates/starter-template/css/style.css"/>
-  <link type="text/css" rel="stylesheet" href="css/stylesheet.css"/>
+  <link type="text/css" rel="stylesheet" href="/assets/materialize.min.css" media="screen,projection"/>
+  <link type="text/css" rel="stylesheet" href="/assets/starter-template.css"/>
+  <link type="text/css" rel="stylesheet" href="/assets/stylesheet.css"/>
   <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
 </head>
 
@@ -96,12 +96,9 @@ spin();</textarea>
     <br />
     <br />
   </footer>
-  
-  <script type="text/javascript" src="bower_components/jquery/dist/jquery.min.js"></script>
-  <script type="text/javascript" src="bower_components/Materialize/dist/js/materialize.js"></script>
-  <script type="text/javascript" src="bower_components/EaselJS/lib/easeljs-0.8.2.combined.js"></script>
-  <script type="text/javascript" src="bower_components/TweenJS/lib/tweenjs-0.6.2.combined.js"></script>
-  <script type="text/javascript" src="bower_components/Materialize/templates/starter-template/js/init.js"></script>
-  <script type="text/javascript" src="src/c-quince/c-quince.js"></script>
+
+  <script src="http://code.createjs.com/easeljs-0.8.2.min.js"></script>
+  <script src="http://code.createjs.com/tweenjs-0.6.0.min.js"></script>
+  <script type="text/javascript" src="/assets/bundle.js"></script>
 </body>
 </html>
