+++ install.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+npm install
+cp node_modules/materialize-css/dist/css/materialize.min.css css/materialize.min.css
+cp node_modules/materialize-css/templates/starter-template/css/style.css css/starter-template.css
